refactor: split camera logic into dedicated hooks
